Move status display styles into a common stylesheet; Refactoring
