+++ src/dashboard/dataImport/main.ts
@@ -1,4 +1,5 @@
-import '../globalStyles.css';
+import '../styles/globalStyles.css';
+import '../styles/statusDisplay.css';
 import './dataImport.css';
 
 import './teamData';
