+++ src/dashboard/rounds/main.ts
@@ -1,6 +1,6 @@
 import { addChangeReminder, fillList } from '../globalScripts';
 
-import '../globalStyles.css';
+import '../styles/globalStyles.css';
 import './rounds.css';
 import { generateId } from '../../helpers/generateId';
 import { ActiveRoundId, Round, Rounds } from 'schemas';
