+++ src/dashboard/settings/main.ts
@@ -1,4 +1,4 @@
-import '../globalStyles.css';
+import '../styles/globalStyles.css';
 import { NodeCGBrowser } from 'nodecg/browser';
 import './lastFmSettings';
 import './radiaSettings';
